Adds intro page to templates documentation (#797)
* first version of intro chapter, typo fixes

* fixes ordering of template articles

* re-adds a few moved sections

* Adds mention of PropChangeEvent
