+++ documentation/polymer-templates/tutorial-template-components-in-slot.asciidoc
@@ -1,13 +1,13 @@
 ---
 title: Using the <slot> Element in a Polymer Template
-order: 4
+order: 11
 layout: page
 ---
 
 
 = Using the <slot> Element in a PolymerTemplate
 
-`PolymerTemplate` cannot not render child templates added using the `Component` or `Element` API. This is because it uses the shadow DOM that renders the *shadow tree*, instead of the element's children that are in the light DOM.
+`PolymerTemplate` cannot render child templates added using the `Component` or `Element` API. This is because it uses the shadow DOM that renders the *shadow tree*, instead of the element's children that are in the light DOM.
 
 To overcome this, you can add a `<slot>` element to mark the place where the light DOM elements should be rendered.
