+++ documentation/polymer-templates/tutorial-template-components.asciidoc
@@ -1,26 +1,20 @@
 ---
 title: Binding Components from a PolymerTemplate
-order: 2
+order: 3
 layout: page
 ---
 
 ifdef::env-github[:outfilesuffix: .asciidoc]
 = Binding Components from a PolymerTemplate
 
-You can add child components to templates using the `Component` or `Element` API.
-[NOTE]
-Added items to the PolymerTemplate only show up on screen if added to the `ShadowRoot` of
-the template if it doesn't have a `<slot>` see. <<tutorial-template-components-in-slot#,Using <slot> in PolymerTemplates>>
-
-By using the `@Id` annotation, you can get a server-side `Component` or `Element`,
-for an element only defined in the PolymerTemplate JavaScript file for the client side to
-use on the server.
+By using the `@Id` annotation, you can get a server-side `Component` or `Element` reference
+for an element defined in the PolymerTemplate JavaScript file. This way you can interact with the `Component` or `Element` on the server side.
 
 [NOTE]
 The `Component`/`Element` needs to have the same `@Tag` as the actual element that is wired.
 This means that you can not wire a `<span id="content"></span>` to a `@Id("content") Div content`
 
-Here is the content of the JavaScript module creating the template with a placeholder `div` element with a `"content"` identifier. This element is mapped to a `Div` component in Java code below to set a `Component` as a child for it.
+Here is a template file with a placeholder `div` element with a `"content"` identifier. This element is mapped to a `Div` component in Java code below so that we can add a `Component` as a child to it.
 
 [source,js]
 ----
@@ -32,7 +26,7 @@ class MainPage extends PolymerElement {
             <div id="content"></div>
             <hr>
             <div id="footer">
-                <a href="mailto:someone@example.com?Subject=Hello%20again" target="_top">Send Mail</a>
+                <a href="mailto:someone@example.com?Subject=Hello" target="_top">Send Mail</a>
             </div>`;
     }
 
@@ -44,7 +38,7 @@ class MainPage extends PolymerElement {
 customElements.define(MainPage.is, MainPage);
 ----
 
-You can implement the method that wires a `Component` to the content of your template element.
+You can for instance implement a method that adds a `Component` to the content of your template element.
 
 [source,java]
 ----
@@ -62,13 +56,13 @@ public class MainPage extends PolymerTemplate<TemplateModel> {
 }
 ----
 
-The `@Id` annotation is used to map a component to an element created by the template
-on the client that has the identifier `"content"`.
+The `@Id` annotation is used to map a component to an element in the template
+on the client that has the HTML identifier `"content"`.
 
 A component instance of the declared type is created automatically and
-wired to the actual dom element and injected into the 'container' field.
+wired to the template DOM element and the `content` field in the Java class.
 
-Now you can set any `Component` as a content for the `MainPage` class:
+You can now call the method to set any `Component` as a content for the `MainPage` class:
 
 [source,java]
 ----
@@ -87,7 +81,7 @@ there is the anchor `a` element available on the client side. The injected insta
 Similar to this the `getText()` method of the `Div` instance injected via
 `@Id("header")` returns an empty string.
 
-So server side `Component`/`Element` read methods are not always in sync
+So, server side `Component`/`Element` read methods are not always in sync
 with the client-side. But you still may use mutation API methods from the server side
 like `appendChild`/`setProperty`/`setAttribute`. Getter methods returns values
 which have been set from the server side only.
@@ -105,4 +99,4 @@ The declared type used in an `@Id` injection declaration must have a default con
 The `@Id` annotation can also be used to inject an `Element` instance instead of a `Component` instance in case you want to use low level API or there is no appropriate HTML component available.
 
 [NOTE]
-You can detect whether a component is part of a Template by using the `isTemplateMapped` method. See the <<tutorial-component-integration#,Integrating components in a PolymerTemplate>> tutorial for more details.
+You can detect whether a component is used in a Template by using the `isTemplateMapped` method. See the <<tutorial-component-integration#,Integrating components in a PolymerTemplate>> tutorial for more details.
