+++ documentation/polymer-templates/tutorial-template-event-handlers.asciidoc
@@ -1,6 +1,6 @@
 ---
 title: Handling User Events in a PolymerTemplate
-order: 5
+order: 4
 layout: page
 ---
 
@@ -8,18 +8,18 @@ ifdef::env-github[:outfilesuffix: .asciidoc]
 = Handling User Events in a PolymerTemplate
 
 == Client Side Event Handlers
-`PolymerTemplate` defines a special syntax `on-_event_="methodName"` for attaching client side event handlers to elements.
-(*Note!* that the method should be without any arguments and have no parenthesis)
+`PolymerTemplate` defines a special syntax `on-_eventname_="methodName"` for attaching client side event handlers to elements.
+(*Note!* that the methodName should be without any arguments and have no parenthesis)
 
 To wire an event to a `PolymerElement` method you can write a template as:
 [source,js]
 ----
-class XCustom extends PolymerElement {
+class EventhandlerDemo extends PolymerElement {
     static get template() {
         return html`<button on-click="handleClick">Say hello</button>`;
     }
 
-    static get is() {return 'x-custom'}
+    static get is() {return 'eventhandler-demo'}
 
     handleClick() {
       console.log('Button was clicked.');
@@ -30,7 +30,7 @@ customElements.define(XCustom.is, XCustom);
 ----
 
 Clicking on the `<button>` will now show an alert in the browser.
-You can listen to any event using the `on-_event_` syntax, it does not matter if it is a built-in browser event or a custom event from e.g. a web component.
+You can listen to any event using the `on-_eventname_` syntax, it does not matter if it is a built-in browser event or a custom event from e.g. a web component. The available events are typically listed in the components API description.
 
 == Server-Side Event Handlers
 To handle a DOM event in a template on the server side you can create a method with the event name and annotate it with `@EventHandler`.
@@ -61,15 +61,14 @@ public class EventHandlerPolymerTemplate extends PolymerTemplate<TemplateModel> 
 }
 ----
 
-The framework will wire up the client-side event when having the `@EventHandler` annotation on the method `handleClick()`.
+The framework will wire up the client-side event when you've added the `@EventHandler` annotation on the method `handleClick()`.
 [NOTE]
-If a case arises where there is a client-side implementation of the server event handler the client side method will
-be executed before the server-side event handler method is called.
+If you add both client and server side event listeners, the client side handlers will be called before the server side handlers.
 
 === Adding Event Data to server-side event
 
 An event can also include additional information about what has happened, e.g. which mouse button was used for a click event.
-When you use `@EventHandler` annotation, all constructor parameters should have an `@EventData` annotation that tells the framework what data to send from the browser.
+When you use the `@EventHandler` annotation, each method parameter should have an `@EventData` annotation that tells the framework what data to send from the browser. You can send any serializeable value present in the client side event object; the available data depends on the type of event.
 
 [source,js]
 ----
@@ -98,15 +97,14 @@ public class EventDataHandlerPolymerTemplate extends PolymerTemplate<TemplateMod
 }
 ----
 
-Now the client would send the extra information back to the server for `event.type`, `event.srcElement.tagName` and the `event.offset[X/Y]` can then be
-used like normal variables.
+The client then parses the event data and sends the extra information back to the server for `event.type`, `event.srcElement.tagName` and the `event.offset[X/Y]`.
 
 [NOTE]
 The server will throw an exception if the `EventData` can not be converted to given format.
 for instance you could get the exception `java.lang.ClassCastException: Cannot cast elemental.json.impl.JreJsonNumber to elemental.json.JsonObject`
 Also the client might throw exceptions if the value given for `EventData` can not be executed or converted to Json.
 
-There is a shorthand for getting model specific item as an object in your event handler. To be able to use it you
+There is a shorthand for getting a Model specific item as an object in your event handler. To be able to use it you
 should define your model (see <<tutorial-template-model-bean#,Using Beans with a PolymerTemplate Model>>) in the template class.
 
 [source,java]
@@ -170,7 +168,8 @@ The method `handleClick` will be called in the server side with the data identif
 You can use the annotation `@ModelItem` with any value provided as a data path.
 By default the data path is `event.model.item`. But your data type should be declared somehow via the model definition (it should be referenced from the model).
 
-[NOTE]
+=== Modifying Model Items before events
+
 Please note that `@ModelItem` is just a convenience way of model data access. The argument which you
 receive in your event handler callback is the *model data* from the server side which you may access directly via your model instance.
 It means that the server doesn't update the model item anyhow from the client. So if you create a custom event on the client side
@@ -232,3 +231,47 @@ The line `this.userInfo.name = this.$.name.value` should be replaced to `this.se
 the correct way to update sub-properties in Polymer. But in this case the server-side model will be updated automatically for
 you and there is no need to send this custom event at all. You may just notify somehow the server about the click event (e.g. via
 `this.$server` and a `@ClientCallable` method, see <<tutorial-template-basic#,Creating A Simple Component Using the Template API>>) and get the model value directly from the server-side model.
+
+=== Other ways of calling the server
+
+Another way to call a server side method is the `@ClientCallable` annotation which allows a Java method to be called from the client side code using the notation `this.$server.serverMethodName(args)`.
+It can be used anywhere in your client side Polymer class implementation. You can pass your own arguments in this methodm, as long as their types matches to method declaration on the server side.
+
+You can also define `PropertyChangeListeners` when working with templates. for more info, please see <<../components/tutorial-enabled-state#enabling-property-changes,Enabling Property Changes>>.
+
+=== Receiving "after server update" event
+
+In some cases you may want to execute some client-side logic after the component
+is updated from the server during a roundtrip.
+E.g. the component constructor is called on the client side but
+this component doesn't yet have data from the server side. So it's too early
+to do anything with the component at this stage.
+In this case you can use the method `afterServerUpdate`. If this method is defined
+for the component it will be called each time after the component is updated
+from the server side, allowing you to configure the component with all data available.
+
+[source,js]
+----
+import {PolymerElement,html} from '@polymer/polymer/polymer-element.js';
+
+class MyComponent extends PolymerElement {
+
+    static get template() {
+        return html`
+            <div>
+                <div>[[text]]</div>
+            </div>`;
+    }
+
+    static get is() {
+          return 'my-component';
+    }
+
+    afterServerUpdate(){
+        console.log("The new 'text' value is: "+this.text);
+    }
+}
+
+customElements.define(MyComponent.is, MyComponent);
+----
+
