+++ documentation/polymer-templates/tutorial-template-model-bean.asciidoc
@@ -1,6 +1,6 @@
 ---
 title: Using Beans with a PolymerTemplate Model
-order: 8
+order: 6
 layout: page
 ---
 
@@ -22,15 +22,15 @@ class MyForm extends PolymerElement {
             <div>
                 <div>
                     <span>First name</span>
-                    <input value="[[person.firstName]]" />
+                    <input value="{{person.firstName}}" />
                 </div>
                 <div>
                     <span>Last name</span>
-                    <input value="[[person.lastName]]" />
+                    <input value="{{person.lastName}}" />
                 </div>
                 <div>
                     <span>Age</span>
-                    <input value="[[person.age]]" />
+                    <input value="{{person.age}}" />
                 </div>
             </div>
             <div>
