+++ documentation/src/main/java/com/vaadin/flow/tutorial/polymer/TemplateInTemplate.java
@@ -17,8 +17,8 @@ package com.vaadin.flow.tutorial.polymer;
 
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.dependency.Uses;
 import com.vaadin.flow.component.polymertemplate.EventHandler;
+import com.vaadin.flow.component.polymertemplate.Id;
 import com.vaadin.flow.component.polymertemplate.PolymerTemplate;
 import com.vaadin.flow.templatemodel.TemplateModel;
 import com.vaadin.flow.tutorial.annotations.CodeFor;
@@ -28,14 +28,14 @@ public class TemplateInTemplate {
 
     @Tag("parent-template")
     @JsModule("./com/example/parent-template.js")
-    @Uses(ChildTemplate.class)
     public class ParentTemplate extends PolymerTemplate<Model> {
+
+        @Id("childTemplate")
+        private ChildTemplate child;
     }
 
     public interface Model extends TemplateModel {
         void setName(String name);
-
-        String getName();
     }
 
     @Tag("child-template")
