+++ documentation/src/main/js/PolymerTemplateEventHandlers.js
@@ -1,12 +1,12 @@
 // tutorial::polymer-templates/tutorial-template-event-handlers.asciidoc
 import {PolymerElement,html} from '@polymer/polymer/polymer-element.js';
 
-class XCustom extends PolymerElement {
+class EventhandlerDemo extends PolymerElement {
     static get template() {
         return html`<button on-click="handleClick">Say hello</button>`;
     }
 
-    static get is() {return 'x-custom'}
+    static get is() {return 'eventhandler-demo'}
 
     handleClick() {
       console.log('Button was clicked.');
@@ -55,3 +55,23 @@ class ContactHandler extends PolymerElement {
     }
 }
 customElements.define(ContactHandler.is, ContactHandler);
+
+class MyComponent extends PolymerElement {
+
+    static get template() {
+        return html`
+            <div>
+                <div>[[text]]</div>
+            </div>`;
+    }
+
+    static get is() {
+          return 'my-component';
+    }
+
+    afterServerUpdate(){
+        console.log("The new 'text' value is: "+this.text);
+    }
+}
+
+customElements.define(MyComponent.is, MyComponent);
