+++ documentation/src/main/js/PolymerTemplateModelBean.js
@@ -7,15 +7,15 @@ class MyForm extends PolymerElement {
             <div>
                 <div>
                     <span>First name</span>
-                    <input value="[[person.firstName]]" />
+                    <input value="{{person.firstName}}" />
                 </div>
                 <div>
                     <span>Last name</span>
-                    <input value="[[person.lastName]]" />
+                    <input value="{{person.lastName}}" />
                 </div>
                 <div>
                     <span>Age</span>
-                    <input value="[[person.age]]" />
+                    <input value="{{person.age}}" />
                 </div>
             </div>
             <div>
